refactor: remove unused pages
