--- frontend/src/components/PaymentForm/index.js
@@ -1,4 +1,0 @@
-import CheckoutForm from './CheckoutForm';
-import WaivingForm from './WaivingForm';
-
-export { CheckoutForm, WaivingForm };
