+++ frontend/src/pages/abstract/submission.js
@@ -1,0 +1,1 @@
+// TODO: there should be a way to request for data removal in this page
