--- frontend/src/pages/academy/volunteer.js
@@ -1,101 +1,0 @@
-import React from 'react';
-import CommonPageStyles from '../../components/BaseComponents/CommonPageStyles';
-import { AcademyLayout } from '../../components/layout';
-
-export default () => (
-  <AcademyLayout>
-    <CommonPageStyles>
-      <h2>Volunteer Recruitment</h2>
-      <p>
-        As Neuromatch Academy rapidly grows, we need volunteers to help us plan
-        and sort out the academy. As a volunteer, you will work directly with the
-        committee chairs to help them manage multiple tasks. Volunteers will be
-        needed in areas such as administration, language issues, communications,
-        and tech. Volunteers will not get payment but we are thrilled to have you
-        contributing and learning with us as we build the academy.
-      </p>
-      <p>
-        Below is a list of currently open positions with contact information.
-        You can contact them directly. We look forward to having you as a part
-        of Neuromatch Academy.
-      </p>
-      <p>
-        <strong>
-          We are still looking for mentors!
-        </strong>
-        We are continuing to accept applications for mentors from faculty, senior postdocs,
-        and members of industry. Mentors help students think through their project
-        plans and offer career advice. Find out more about the position and
-        apply
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          href="https://semtsinghua.au1.qualtrics.com/jfe/form/SV_eVfkK52NATb06ah"
-        >
-          here
-        </a>
-        .
-      </p>
-      <h3>
-        Front-end developer.
-      </h3>
-      <p>
-        <b>Commitment</b>
-        : ~3-8 hours per week.
-        <br />
-        <b>Contact</b>
-        : Patrick Mineault (
-        <code>patrick.mineault@gmail.com</code>
-        )
-        <br />
-        <b>Job Description</b>
-        : You will help continue to develop the front-end of the Neuromatch Academy web page.
-        <br />
-        <b>Qualifications</b>
-        : The ability to work with JavaScript and CSS (managing a website hosted or developed
-        on GitHub with Vercel and Gatsby a plus).
-      </p>
-      <h3>
-        Video editor.
-      </h3>
-      <p>
-        <b>Commitment</b>
-        : ~4-8 hours per week.
-        <br />
-        <b>Contact</b>
-        : Patrick Mineault (
-        <code>patrick.mineault@gmail.com</code>
-        )
-        <br />
-        <b>Job Description</b>
-        : You will edit and polish lecture videos so students have a great experience.
-        You&apos;ll suggest improvements in recording setup to improve capture quality. We&apos;ll
-        help you get started with the necessary tools.
-        <br />
-        <b>Qualifications</b>
-        : Experience with editing videos or willingness to be trained in video and audio
-        editing software.
-      </p>
-      <h3>
-        Caption editor.
-      </h3>
-      <p>
-        <b>Commitment</b>
-        : ~4-8 hours per week.
-        <br />
-        <b>Contact</b>
-        : Brad Wyble (
-        <code>bwyble@gmail.com</code>
-        )
-        <br />
-        <b>Job Description</b>
-        : You will edit and polish automatically generated close captioning in tight
-        coordination with our video editing team to increase the ability of students
-        to understand our content. We&apos;ll help you get started with the necessary tools.
-        <br />
-        <b>Qualifications</b>
-        : Basic fluency with computational and neuroscience jargon in English.
-      </p>
-    </CommonPageStyles>
-  </AcademyLayout>
-);
