+++ frontend/src/pages/faq.js
@@ -10,6 +10,8 @@ const StyledIframe = styled.iframe`
   margin-top: 15px;
 `;
 
+// change to yaml
+
 export default () => (
   <Layout noPadding>
     <Helmet>
